Java8 module which will use the native long adder impl
