+++ metrics-core/src/main/java/io/dropwizard/metrics/EWMA.java
@@ -29,7 +29,7 @@ public class EWMA {
     private volatile boolean initialized = false;
     private volatile double rate = 0.0;
 
-    private final LongAdder uncounted = new LongAdder();
+    private final LongAdder uncounted = new LongAdderImpl();
     private final double alpha, interval;
 
     /**
