+++ metrics-core/src/main/java/io/dropwizard/metrics/Meter.java
@@ -22,7 +22,7 @@ public class Meter implements Metered {
     private final EWMA m5Rate = EWMA.fiveMinuteEWMA();
     private final EWMA m15Rate = EWMA.fifteenMinuteEWMA();
 
-    private final LongAdder count = new LongAdder();
+    private final LongAdder count = new LongAdderImpl();
     private final long startTime;
     private final AtomicLong lastTick;
     private final Clock clock;
